# cmd/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Service for managing and sending notifications (Take Home Challenge)",
        "title": "Notification Manager System - THC",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Julián Rentería",
            "email": "imakheri@gmail.com"
        },
        "version": "0.1"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/user": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Allows the authenticated user to update their information (except email)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Update user data",
                "parameters": [
                    {
                        "description": "User information for update",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.UpdateUserDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "user updated successfully",
                        "schema": {
                            "$ref": "#/definitions/dtos.ResponseUserDTO"
                        }
                    },
                    "400": {
                        "description": "invalid user properties",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponseDTO"
                        }
                    },
                    "404": {
                        "description": "record not found",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponseDTO"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponseDTO"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Authenticates an user and return thier data along a JWToken",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Loggin / Get user data",
                "parameters": [
                    {
                        "description": "Access credentials",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.GetUserDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "user logged in successfully",
                        "schema": {
                            "$ref": "#/definitions/dtos.LoginResponseDTO"
                        }
                    },
                    "400": {
                        "description": "invalid user credentials",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponseDTO"
                        }
                    },
                    "401": {
                        "description": "the e-mail address or password is incorrect",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponseDTO"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponseDTO"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dtos.ErrorResponseDTO": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "dtos.GetUserDTO": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "williams@example.com"
                },
                "password": {
                    "type": "string",
                    "example": "Hamlet123."
                }
            }
        },
        "dtos.LoginResponseDTO": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/dtos.ResponseUserDTO"
                }
            }
        },
...
        "dtos.ResponseUserDTO": {
            "type": "object",
            "properties": {
                "device_token": {
                    "type": "string",
                    "example": "0q1e2r3t4y5a6b7c8d9e"
                },
                "email": {
                    "type": "string",
                    "example": "williams@example.com"
                },
                "name": {
                    "type": "string",
                    "example": "William Shakespeare"
                },
                "phone": {
                    "type": "string",
                    "example": "1234567890"
                }
            }
        },
...
        "dtos.UpdateUserDTO": {
            "type": "object",
            "properties": {
                "device_token": {
                    "type": "string",
                    "example": "12345qwerty67890"
                },
                "name": {
                    "type": "string",
                    "example": "Fernando Pessoa"
                },
                "password": {
                    "type": "string",
                    "example": "Mensagem123."
                },
                "phone": {
                    "type": "string",
                    "example": "0129834765"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "api_key",
            "in": "header"
        },
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}